backend/requirements.txt
# Backend dependencies for AI Backtest Platform

# FastAPI and web server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Celery and Redis
celery>=5.4.0
redis>=5.2.0

# SSE support
sse-starlette>=2.1.0

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client (for testing)
httpx>=0.27.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Engine dependencies (also needed for subprocess)
pandas>=2.2.0
numpy>=1.26.0

# Performance optimization
numba>=0.60.0
